live N, H20, start P2O5

+++ app/src/main/java/com/example/yandre/biontest/database/model/VodopotrebModel.java
@@ -9,9 +9,9 @@ public class VodopotrebModel {
     @PrimaryKey(autoGenerate = true)
     private long id;
     private String culture;
-    private int value;
+    private double value;
 
-    public VodopotrebModel(String culture, int value) {
+    public VodopotrebModel(String culture, double value) {
         this.culture = culture;
         this.value = value;
     }
@@ -32,11 +32,11 @@ public class VodopotrebModel {
         this.culture = culture;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
 }
